requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
lxml==4.9.3
openai==1.3.5
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyPDF2==3.0.1
pypdfium2==4.24.0
python-docx==1.1.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
